修正

+++ src/main/java/com/yicloud/trans/model/mssql/Jbxxk.java
@@ -14,7 +14,7 @@ import lombok.Data;
 @Data
 public class Jbxxk implements Serializable {
     @TableId
-    private Integer zyh;
+    private Long zyh;
 
     private String xm;
